requirements.txt
fastapi
uvicorn
pydantic-settings
pyyaml
requests
